Cleanup. Move test converter to tests folder.
