+++ stablehlo/dialect/CMakeLists.txt
@@ -130,7 +130,6 @@ add_mlir_dialect_library(StablehloOps
   LINK_LIBS PUBLIC
   StablehloAssemblyFormat
   StablehloBase
-  StablehloCompatibility
   StablehloTypeInference
   MLIRArithDialect
   MLIRDataLayoutInterfaces
