+++ stablehlo/tests/ops_chlo_roundtrip.mlir
@@ -2,7 +2,7 @@
 // RUN: stablehlo-opt -emit-bytecode %s | stablehlo-opt | FileCheck %s
 // RUN: stablehlo-opt -emit-bytecode -debug-only=chlo-bytecode %s 2>&1 | (! grep 'Not Implemented')
 // RUN: stablehlo-opt -emit-bytecode %s | stablehlo-opt -debug-only=chlo-bytecode 2>&1 | (! grep 'Not Implemented')
-// RUN: stablehlo-translate -compat --emit-bytecode %s | stablehlo-translate --compat --target=999
+// RUN: stablehlo-translate -compat %s | stablehlo-translate --compat --emit-assembly --target=999
 
 
 module {
